Refactor agent into modular components
