+++ IMO25/code/config.py
@@ -1,0 +1,14 @@
+from __future__ import annotations
+import os
+from dotenv import load_dotenv
+
+load_dotenv()
+
+STRATEGIST_MODEL_NAME = os.getenv("STRATEGIST_MODEL_NAME", "deepseek/deepseek-r1-0528:free")
+WORKER_MODEL_NAME = os.getenv("WORKER_MODEL_NAME", "deepseek/deepseek-chat-v3-0324:free")
+IMPROVER_MODEL_NAME = os.getenv("IMPROVER_MODEL_NAME", "deepseek/deepseek-chat-v3-0324:free")
+ENABLE_NRPA = os.getenv("NRPA_ENABLED", "1") in ("1", "true", "True", "yes", "YES")
+
+API_URL_BASE = "https://openrouter.ai/api/v1/chat/completions"
+MODEL_PROVIDER = os.getenv("MODEL_PROVIDER", "openrouter").lower()
+CEREBRAS_MODEL_DEFAULT = os.getenv("CEREBRAS_MODEL_DEFAULT", "llama-4-scout-17b-16e-instruct")
